[Testing]: Add testing infrastructure and CI pipelines (#52)
- Integrate testcoe via CMake with optional GAMECOE_USE_TESTCOE flag
- Create module-based test structure (tests/entity/, tests/integration/)
- Implement custom test runner with CLI options (--help, --all, --suite, --test)
- Add tests across 4 suites (EntityTests, SparseSetTests, ComponentPoolTests, ECSIntegrationTests)
- Add CI workflows for Linux (GCC/Clang), macOS (Clang), Windows (MSVC/MinGW)
- Rename old architecture files to .old (isolate during DoD ECS refactoring)
- Delete empty files (material, mesh, sprite)

+++ src/inputcoe.cpp
@@ -1,5 +1,4 @@
 #include <inputcoe.hpp>
-#include <gamecoe/core/game.hpp>
 #include <GLFW/glfw3.h>
 #include <unordered_map>
 #include <gamecoe_config.hpp>
@@ -96,14 +95,14 @@ namespace inputcoe
     {
         void mousePositionCallback(GLFWwindow *window, double xpos, double ypos)
         {
-            g_currentMousePosition[0] = xpos;
-            g_currentMousePosition[1] = ypos;
+            g_currentMousePosition[0] = static_cast<float>(xpos);
+            g_currentMousePosition[1] = static_cast<float>(ypos);
 
             int width, height;
             glfwGetFramebufferSize(window, &width, &height);
 
-            g_currentMousePositionNormalized[0] = xpos / width;
-            g_currentMousePositionNormalized[1] = ypos / height;
+            g_currentMousePositionNormalized[0] = static_cast<float>(xpos) / width;
+            g_currentMousePositionNormalized[1] = static_cast<float>(ypos) / height;
         }
 
         void keyCallback([[maybe_unused]] GLFWwindow *window, int key, [[maybe_unused]] int scancode, int action, [[maybe_unused]] int mods)
